fix: clarify time-ordering features of v1 UUIDs
The analysis text could be misunderstood in a sense that all v1 UUIDs
are globally time ordered. This is not the case, since the timestamp is
split into three fields that appear in the UUIDs in reverse order, see
https://github.com/kelektiv/node-uuid/issues/75#issuecomment-25893270
for more discussion.

This improvement of the text does not change anything about the
assumptions or results presented in the analysis.