feat: remove section about treating v4 as a default
Discussion in #3 showed that there is general agreement for only
supporting the v4 algorithm.

However, there are concerns that, if we design an API that promotes one
algorithm as the default, this assumption might not hold in the future.
In order to provide a more future-proof API we may therefore only
support certain algorithms, but will likely not treat any of the
supported algorithms in a speacial way.